+++ aura/__init__.py
@@ -1,3 +1,3 @@
 """Aura language toolchain."""
 
-__version__ = "0.2.0a3"
+__version__ = "0.2.0a4"